extract merge action from taskControls to its parent component
